requirements.txt
Django==3.2.18
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2 
djoser==2.1.0
drf-yasg2==1.19.4
python-dotenv==0.20.0